+++ app/src/main/java/com/example/proyectoglucapp/SharedViewModel.kt
@@ -3,9 +3,9 @@ package com.example.proyectoglucapp
 import androidx.lifecycle.ViewModel
 
 class SharedViewModel : ViewModel() {
-    var sensibilidad: Float? = null
-    var ratioManana: Float? = null
-    var ratioMediodia: Float? = null
-    var ratioTarde: Float? = null
-    var ratioNoche: Float? = null
+    var sensibilidad: Double? = null
+    var ratioManana: Double? = null
+    var ratioMediodia: Double? = null
+    var ratioTarde: Double? = null
+    var ratioNoche: Double? = null
 }
